requirements.txt
streamlit
fastapi
sqlalchemy
pyjwt
plotly
pandas
uvicorn
python-multipart
numpy
requests